:truck: map 관련 ui 파일 이동 #68
