--- src/models/map/map.d.ts
@@ -1,1 +1,0 @@
-type MapId = keyof typeof import('@/models/map/index').MAPS;
