+++ src/models/map/types.d.ts
@@ -1,0 +1,1 @@
+type MapId = keyof typeof import('./constants').MAPS;
